fix(backend): clear the mypy and naming gates the release workflow enforces
CI runs `uv run mypy app/` and `npm run check:naming` before it builds, and
both were red.

auto_layout is annotated to return ints but centres a column's rows on a
half-cell, so it returns floats — which is what Component.col/row already are.
Widened the annotation rather than truncating the layout. The infix-word test in
peek() narrowed on truthiness, which mypy cannot follow through to .lower().

Two camelCase service parameters, one introduced here and one predating it, are
renamed to snake_case. Only the Python names move: `fanoutId` stays the wire key
in the SSE payloads and the fanout snapshot, which is the dual naming the docs
already describe.

+++ backend-py/app/services/tools/schematic.py
@@ -162,7 +162,7 @@ _COL_STEP = 5
 _ROW_STEP = 6
 
 
-def auto_layout(elements: list[Element]) -> dict[str, dict[str, int]]:
+def auto_layout(elements: list[Element]) -> dict[str, dict[str, float]]:
     """Place parts left-to-right by signal depth, stacking parallel branches
     into rows.
 
@@ -220,7 +220,7 @@ def auto_layout(elements: list[Element]) -> dict[str, dict[str, int]]:
     for ref in refs:  # keep deck order for determinism
         per_column.setdefault(depth[ref], []).append(ref)
 
-    layout: dict[str, dict[str, int]] = {}
+    layout: dict[str, dict[str, float]] = {}
     for col_index, refs_in_col in sorted(per_column.items()):
         span = len(refs_in_col)
         for slot, ref in enumerate(refs_in_col):
